VlasovSolvers.jl Documentation

First draft of a Vlasov solvers Julia suite.

Backward Semi-Lagrangian

Pseudo-Spectral

Particle-In-Cell

Types and functions

VlasovSolvers.advection! — Method

advection_v!(f, dt)

function to advect the distribution function f with velocity v during a time step dt. Interpolation method uses bspline periodic.

VlasovSolvers.bspline — Method

bspline(p, j, x)

Return the value at x in [0,1[ of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

VlasovSolvers.compute_e — Method

compute_e(f)

compute Ex using that -ik*Ex = rho

VlasovSolvers.compute_rho — Method

compute_rho(f)

Compute charge density ρ(x,t) = ∫ f(x,v,t) dv