VlasovSolvers.jl Documentation

First draft of a Vlasov solver.

Vlasov-Poisson equation

We consider the dimensionless Vlasov-Poisson equation for one species with a neutralizing background.

\[\frac{\partial f}{\partial t}+ v\cdot \nabla_x f + E(t,x) \cdot \nabla_v f = 0,\]

\[- \Delta \phi = 1 - \rho, E = - \nabla \phi\]

\[\rho(t,x) = \int f(t,x,v)dv.\]

Vlasov Equation - Wikipedia

Input parameters

using VlasovSolvers

dev = CPU()                  # device
nx, nv = 64, 64              # grid resolution
stepper = StrangSplitting()  # timestepper
dt = 0.1                     # timestep
nsteps = 1000                # total number of time-steps

xmin, xmax = 0, 4π           # X Domain length (m)
vmin, vmax = -6, 6           # V Domain length (m)
α  = 0.001                   # Perturbation amplitude
kx = 0.5                     # Wave number of perturbation

xgrid = OneDGrid(dev, nx, xmin, xmax)
vgrid = OneDGrid(dev, nv, vmin, vmax)

f = DistributionFunction( xgrid, vgrid )

landau!(f, α, kx)

prob = VlasovProblem(f, BSL(5), dev)

sol = solve!(prob, stepper, dt, nsteps )

t  = range(0.0, stop=tf, length=nt)

plot( t, sol; label = "E")
plot!(t, -0.1533*t.-5.50; label="-0.1533t.-5.5")

Types and functions

VlasovSolvers.advection! — Method

advection_v!(f, dt)

function to advect the distribution function f with velocity v during a time step dt. Interpolation method uses bspline periodic.

VlasovSolvers.bspline — Method

bspline(p, j, x)

Return the value at x in [0,1[ of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

VlasovSolvers.compute_e — Method

compute_e(f)

compute Ex using that -ik*Ex = rho

VlasovSolvers.compute_rho — Method

compute_rho(f)

Compute charge density ρ(x,t) = ∫ f(x,v,t) dv